pyproject.toml
[tool.poetry]
name = "library-drf"
version = "0.1.0"
description = ""
authors = ["Pvl1307 <ivanov.pavel.1307@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
redis = "^5.0.1"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
